Removed SparkListenerRDDCreation event
In the original design of Arthur, RDD creation event is reported by all RDD instances at the end of their constructor.  The drawback of this approach is that no RDD class can be ever inherited.  Otherwise, once the sub class is instantiated, there would be two RDD creation events reported, one for the sub class, one for the super class.

Further more, in the original design, RDDs that are parents of other RDDs are serialized multiple times because once an RDD is serialized, their dependencies are serialized together.

This commit removed the RDD creation event.  Instead, RDDs are collected from the ActiveJob object reported together with the SparkListenerJobStart event.  It means that, RDD are not recored into the event log if no related job is started.